+++ re2/testing/backtrack.cc
@@ -230,7 +230,7 @@ bool Backtracker::Try(int id, const char* p) {
           (longest_ && p > submatch_[0].end())) {  // ... or better match
         for (int i = 0; i < nsubmatch_; i++)
           submatch_[i].set(cap_[2*i],
-                           static_cast<int>(cap_[2*i+1] - cap_[2*i]));
+                           static_cast<size_t>(cap_[2*i+1] - cap_[2*i]));
       }
       return true;
